requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
onnxruntime==1.17.3
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
